+++ src/components/admin/uploadMP/formUploader/formUploaderResource.ts
@@ -1,0 +1,91 @@
+export const 정당리스트 = [
+  "더불어민주당",
+  "국민의힘",
+  "정의당",
+  "진보당",
+  "기본소득당",
+  "시대전환",
+  "한국의희망",
+  "무소속",
+];
+
+export const 상임위원회리스트 = [
+  "국회운영위원회",
+  "법제사법위원회",
+  "정무위원회",
+  "기획재정위원회",
+  "교육위원회",
+  "과학기술정보방송통신위원회",
+  "외교통일위원회",
+  "국방위원회",
+  "행정안전위원회",
+  "문화체육관광위원회",
+  "농림축산식품해양수산위원회",
+  "산업통상자원중소벤처기업위원회",
+  "환경노동위원회",
+  "국토교통위원회",
+  "정보위원회",
+  "여성가족위원회",
+  "보건복지위원회",
+];
+
+export const 지역구리스트 = [
+  "서울특별시",
+  "부산광역시",
+  "인천광역시",
+  "대구광역시",
+  "광주광역시",
+  "대전광역시",
+  "울산광역시",
+  "세종특별자치시",
+  "경기",
+  "강원",
+  "충북",
+  "충남",
+  "전북",
+  "전남",
+  "경북",
+  "경남",
+  "제주",
+];
+
+export const 분구리스트 = ["갑", "을", "병", "정", "무"];
+
+export const 공약이행현황 = {
+  title: "공약 이행 현황",
+  subtitle:
+    "총 공약수 = 완료 + 추진 중 + 보류 + 폐기 + 기타 공약 수 (합계가 일치해야 함)",
+  theadList: ["총공약", "완료", "추진중", "보류", "폐기", "기타"],
+  tbodyList: ["공약수"],
+};
+export const 성격내용별완료현황 = {
+  title: "성격·내용별 공약 이행 완료 현황",
+  subtitle:
+    "총 공약수 = 완료 + 추진 중 + 보류 + 폐기 + 기타 공약 수 (합계가 일치해야 함)",
+  theadList: [
+    "국정공약",
+    "지역공약",
+    "입법공약",
+    "재정공약",
+    "임기내",
+    "임기후",
+    "지속사업",
+    "신규사업",
+  ],
+  tbodyList: ["완료 공약수", "전체 공약수"],
+};
+
+export const 입법현황 = {
+  title: "입법 현황",
+  subtitle:
+    "필요입법 공약 총 수: 입법이 필요한 공약의 총 수,\n입법 의결 완료 공약 총 수: 입법을 모두 완료(의결)한 공약의 총 수",
+  theadList: ["필요입법 공약 총 수", "입법 의결 완료 공약 총 수"],
+  tbodyList: ["공약수"],
+};
+export const 재정현황 = {
+  title: "재정 현황",
+  subtitle: "전체 공약의 재정 현황 합계",
+  theadList: ["필요재정 총액", "확보재정 총액", "집행재정 총액"],
+  tbodyList: ["금액"],
+  unit: "원",
+};
